Create Appointment component for use in provider and staff appointment
